+++ src/main/java/com/homedepot/mm/po/allocationteamdata/entities/teradata/BayParm.java
@@ -1,5 +1,7 @@
 package com.homedepot.mm.po.allocationteamdata.entities.teradata;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.Table;
@@ -21,6 +23,6 @@ public class BayParm {
 	private int uploadid;
 	private String locationid;
 	private String productcode;
-	private double bay_parm_val;
+	private BigDecimal bay_parm_val;
 	private String activeflag;
 }
